reverting v2beta1, support older API endpoints for now
