+++ cmd/cluster-agent/api/server.go
@@ -121,8 +121,8 @@ func validateToken(next http.Handler) http.Handler {
 func isExternalPath(path string) bool {
 	if strings.HasPrefix(path, "/api/v1/metadata/") && len(strings.Split(path, "/")) == 7 || // support for agents < 6.5.0
 		path == "/version" ||
-		strings.HasPrefix(path, "/api/v2beta1/tags/pods/") && len(strings.Split(path, "/")) == 8 ||
-		strings.HasPrefix(path, "/api/v2beta1/tags/nodes/") && len(strings.Split(path, "/")) == 6 {
+		strings.HasPrefix(path, "/api/v1/tags/pods/") && len(strings.Split(path, "/")) == 8 ||
+		strings.HasPrefix(path, "/api/v1/tags/nodes/") && len(strings.Split(path, "/")) == 6 {
 		return true
 	}
 	return false
